remove luxor dependency
